Temporary remove meta_call
