+++ frame/pallet-account/src/tests.rs
@@ -18,7 +18,7 @@
 
 use super::*;
 use assert_matches::assert_matches;
-use frame_support::{assert_err, assert_ok};
+use frame_support::assert_ok;
 use mock::*;
 
 #[test]
@@ -78,63 +78,6 @@ pub fn proxy_call_works() {
                 .clone(),
             RuntimeEvent::Account(Event::ProxyCall{origin, ..})
             if origin == NativeAndEVM::Native(ALICE_D1_NATIVE.into())
-        );
-    })
-}
-
-#[test]
-pub fn meta_call_works() {
-    use parity_scale_codec::Encode;
-    use sp_core::Pair;
-    use sp_runtime::traits::IdentifyAccount;
-
-    ExternalityBuilder::build().execute_with(|| {
-        let call: RuntimeCall = pallet_balances::Call::transfer {
-            dest: BOB,
-            value: 10,
-        }
-        .into();
-
-        let pair = sp_core::ed25519::Pair::from_string("//Alice", None).unwrap();
-        let payload = (ChainMagic::get(), 0u64, call.clone());
-        let signer: sp_runtime::MultiSigner = pair.public().into();
-        let account_id = signer.into_account();
-        let signature = pair.sign(&payload.encode()[..]);
-
-        // Make call with signer as origin
-        assert_eq!(System::account(&account_id).nonce, 0);
-
-        assert_ok!(Account::meta_call(
-            RuntimeOrigin::signed(ALICE).into(),
-            Box::new(call),
-            account_id.clone(),
-            signature.into(),
-        ));
-
-        assert_eq!(System::account(&account_id).nonce, 1);
-        assert_eq!(System::account(BOB).data.free, 810);
-    })
-}
-
-#[test]
-pub fn meta_call_bad_signature() {
-    ExternalityBuilder::build().execute_with(|| {
-        let call: RuntimeCall = pallet_balances::Call::transfer {
-            dest: BOB,
-            value: 10,
-        }
-        .into();
-
-        let bad_signature = sp_runtime::MultiSignature::Ecdsa(Default::default());
-
-        assert_err!(
-            Account::meta_call(
-                RuntimeOrigin::signed(ALICE).into(),
-                Box::new(call),
-                ALICE,
-                bad_signature,
-            ),
-            Error::<TestRuntime>::BadSignature,
         );
     })
 }
